+++ game/bagtiles.py
@@ -1,0 +1,132 @@
+#FALTA: omit C+H instead of CH/2xL instead of LL/2xR instead of RR
+#joker aparte porque  tiene puntaje variable
+import random
+class Tile:
+    def __init__(self, letter, points):
+        self.letter = letter
+        self.points = points
+
+class joker(Tile):
+    def __init__(self):
+        super().__init__('joker', 0)
+
+    def select_letter(self, selection):
+        selection = selection.upper()
+        for tile in self.tiles:
+            if selection == tile.letter:
+                self.letter = selection
+                self.points = tile.points
+        return True
+
+class TileBag:
+    def __init__(self):
+        self.tiles = [
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('A', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('E', 1),
+            Tile('I', 1),
+            Tile('I', 1),
+            Tile('I', 1),
+            Tile('I', 1),
+            Tile('I', 1),
+            Tile('I', 1),
+            Tile('L', 1),
+            Tile('L', 1),
+            Tile('L', 1),
+            Tile('L', 1),
+            Tile('N', 1),
+            Tile('N', 1),
+            Tile('N', 1),
+            Tile('N', 1),
+            Tile('N', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('O', 1),
+            Tile('R', 1),
+            Tile('R', 1),
+            Tile('R', 1),
+            Tile('R', 1),
+            Tile('R', 1),
+            Tile('S', 1),
+            Tile('S', 1),
+            Tile('S', 1),
+            Tile('S', 1),
+            Tile('S', 1),
+            Tile('S', 1),
+            Tile('T', 1),
+            Tile('T', 1),
+            Tile('T', 1),
+            Tile('T', 1),
+            Tile('U', 1),
+            Tile('U', 1),
+            Tile('U', 1),
+            Tile('U', 1),
+            Tile('U', 1),
+            Tile('D', 2),
+            Tile('D', 2),
+            Tile('D', 2),
+            Tile('D', 2),
+            Tile('D', 2),
+            Tile('G', 2),
+            Tile('G', 2),
+            Tile('B', 3),
+            Tile('B', 3),
+            Tile('C', 3),
+            Tile('C', 3),
+            Tile('C', 3),
+            Tile('C', 3),
+            Tile('M', 3),
+            Tile('M', 3),
+            Tile('P', 3),
+            Tile('P', 3),
+            Tile('F', 4),
+            Tile('H', 4),
+            Tile('H', 4),
+            Tile('V', 4),
+            Tile('Y', 4),
+            Tile('CH',5),
+            Tile('Q', 5),
+            Tile('J', 8),
+            Tile('LL',8),
+            Tile('Ñ', 8),
+            Tile('RR',8),
+            Tile('X', 8),
+            Tile('Z',10),
+        ]
+        random.shuffle(self.tiles)
+
+    def take(self, count):
+        tiles = []
+        for _ in range(count):
+            tiles.append(self.tiles.pop())
+        return tiles
+
+    def put(self, tiles):
+        self.tiles.extend(tiles)