+++ game/cli.py
@@ -1,0 +1,8 @@
+#interfaz visual
+from game.scrabble import ScrabbleGame
+
+def main():
+    pass
+
+if __name__ == '__main__':
+    main()