+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+##[0.0.5] ~ 03-09-2023
+### Added
+- Module Board.
+- Methods: is empty, calculate word value, validate word in/out board(horizontally and vertically).
+- Module Cell.
+- Methods: WIP.
+- Module Scrabble.
+- Methods: WIP.
+- Module Main.
+- Methods: WIP.
+
+### Changed
+- Module Tile added joker tile.
+
+##[0.0.4] ~ 27-08-2023
+### Added
+- Various modules to complete the bare game structure.
+- A main file to run the game(by now the game gives the player a selection of tiles).
+
+##[0.0.3] ~ 24-08-2023
+### Added
+- Module Player.
+- The player now has a list of letters.
+- The player can now take letters from the bag.
+
+### Changed
+- The dictionary is now a list of words.
+
 ##[0.0.2] ~ 20-08-2023
 ### Added
 - Dictionary with each value and how many of them there are based on Scrable international spanish(100 words).
@@ -9,7 +37,6 @@ All notable changes to this project will be documented in this file.
 - Small changes to the tests.
 
 ##[0.0.1] ~ 18-08-2023
-
 ### Added
 - Wip version of the game.
 - README file with a short description of the game.